% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CGIloaddercompare}
\alias{CGIloaddercompare}
\title{CGIloaddercompare}
\usage{
CGIloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
CGIloaddercompare
}
