% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CGIloaddercompareOpt}
\alias{CGIloaddercompareOpt}
\title{CGIloaddercompareOpt}
\usage{
CGIloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
CGIloaddercompareOpt
}
