% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CLINGENloaddercompare}
\alias{CLINGENloaddercompare}
\title{CLINGENloaddercompare}
\usage{
CLINGENloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
CLINGENloaddercompare
}
