% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CLINGENloaddercompareOpt}
\alias{CLINGENloaddercompareOpt}
\title{CLINGENloaddercompareOpt}
\usage{
CLINGENloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
CLINGENloaddercompareOpt
}
