% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{convertgenesymbolmessage}
\alias{convertgenesymbolmessage}
\title{convertgenesymbolmessage}
\usage{
convertgenesymbolmessage(genes, organism = "human")
}
\arguments{
\item{genes}{A vector with gene symbol.}

\item{organism}{A string parameter: "human" if you have human genes and "mouse if you have mouse genes}
}
\description{
This function give a message with the homologues map
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
