% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CRBloaddercompare}
\alias{CRBloaddercompare}
\title{CRBloaddercompare}
\usage{
CRBloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
CRBloaddercompare
}
