% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CRBloaddercompareOpt}
\alias{CRBloaddercompareOpt}
\title{CRBloaddercompareOpt}
\usage{
CRBloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
CRBloaddercompareOpt
}
