% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CTDloaddercompare}
\alias{CTDloaddercompare}
\title{CTDloaddercompare}
\usage{
CTDloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
CTDloaddercompare
}
