% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{CTDloaddercompareOpt}
\alias{CTDloaddercompareOpt}
\title{CTDloaddercompareOpt}
\usage{
CTDloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
CTDloaddercompareOpt
}
