% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{entrezmap}
\alias{entrezmap}
\title{entrezmap}
\usage{
entrezmap(genes, human = T)
}
\arguments{
\item{genes}{A vector with gene symbol.}

\item{human}{A logical parameter if you have human gene symbol}
}
\description{
This function give a message with the entrez map
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
