% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{findoverlapgenes}
\alias{findoverlapgenes}
\title{findoverlapgenes}
\usage{
findoverlapgenes(phenos)
}
\arguments{
\item{phenos}{A vector with the phenotypes id that you want.}
}
\description{
This function get the genes that are linked to the phenotypes selected.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
