% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{FindPhenoFromGenes}
\alias{FindPhenoFromGenes}
\title{FindPhenoFromGenes}
\usage{
FindPhenoFromGenes(genes, phenoid, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with genes symbol.}

\item{phenoid}{A vector with id terms to get.}

\item{database}{A string parameter with different options:  "HPO", "DIS", "CRB" and "MGD".}

\item{organism}{A string parameter: "human" if you have human genes symbol and "mouse" if you have mouse genes symbol}
}
\description{
Get the genes selected and only their phenotypical terms defined by the user.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
