% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{foundgeneoverlap}
\alias{foundgeneoverlap}
\title{foundgeneoverlap}
\usage{
foundgeneoverlap(geneset, genesetcompare)
}
\arguments{
\item{geneset}{A vector with human gene symbol.}

\item{genesetcompare}{A vector with human gene symbol.}
}
\description{
This function give no overlapping genes
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
