% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{GENOMICS_ENGLANDloadder}
\alias{GENOMICS_ENGLANDloadder}
\title{GENOMICS_ENGLANDloadder}
\usage{
GENOMICS_ENGLANDloadder(genes, organism = "human", plotnumber = 30, url = F)
}
\description{
GENOMICS_ENGLANDloadder
}
