% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{GENOMICS_ENGLANDloaddercompare}
\alias{GENOMICS_ENGLANDloaddercompare}
\title{GENOMICS_ENGLANDloaddercompare}
\usage{
GENOMICS_ENGLANDloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
GENOMICS_ENGLANDloaddercompare
}
