% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{GENOMICS_ENGLANDloaddercompareOpt}
\alias{GENOMICS_ENGLANDloaddercompareOpt}
\title{GENOMICS_ENGLANDloaddercompareOpt}
\usage{
GENOMICS_ENGLANDloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
GENOMICS_ENGLANDloaddercompareOpt
}
