% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{GENOMICS_ENGLANDloadderOpt}
\alias{GENOMICS_ENGLANDloadderOpt}
\title{GENOMICS_ENGLANDloadderOpt}
\usage{
GENOMICS_ENGLANDloadderOpt(genes, organism = "human", plotnumber = 30)
}
\description{
GENOMICS_ENGLANDloadderOpt
}
