% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{getdbnames}
\alias{getdbnames}
\title{getdbnames}
\usage{
getdbnames()
}
\description{
This function show the databases available in PhenoExam package.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
