% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{HPOloaddercompare}
\alias{HPOloaddercompare}
\title{HPOloaddercompare}
\usage{
HPOloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
HPOloaddercompare
}
