% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{HPOloaddercompareOpt}
\alias{HPOloaddercompareOpt}
\title{HPOloaddercompareOpt}
\usage{
HPOloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
HPOloaddercompareOpt
}
