% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{ListGenesPheno}
\alias{ListGenesPheno}
\title{ListGenesPheno}
\usage{
ListGenesPheno(genes, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with the genes symbol}

\item{database}{A string parameter with different options. You can define a vector with the databases.}

\item{organism}{A string parameter: "human" if you have human genes symbol and "mouse" if you have mouse genes symbol}
}
\description{
This function make a list with genes selected and his phenotypical terms
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
