% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{MGDloaddercompare}
\alias{MGDloaddercompare}
\title{MGDloaddercompare}
\usage{
MGDloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
MGDloaddercompare
}
