% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{MGDloaddercompareOpt}
\alias{MGDloaddercompareOpt}
\title{MGDloaddercompareOpt}
\usage{
MGDloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
MGDloaddercompareOpt
}
