% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{ORPHANETloaddercompare}
\alias{ORPHANETloaddercompare}
\title{ORPHANETloaddercompare}
\usage{
ORPHANETloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
ORPHANETloaddercompare
}
