% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{ORPHANETloaddercompareOpt}
\alias{ORPHANETloaddercompareOpt}
\title{ORPHANETloaddercompareOpt}
\usage{
ORPHANETloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
ORPHANETloaddercompareOpt
}
