% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PhenoEnrichCompare}
\alias{PhenoEnrichCompare}
\title{PhenoEnrichCompare}
\usage{
PhenoEnrichCompare(
  geneset,
  genesetcompare,
  database = "HPO",
  organism = "human",
  plotn = 30,
  plotnumber = 1:plotn,
  url = F
)
}
\arguments{
\item{database}{A string parameter with different options. You can define a vector with the databases.}

\item{organism}{A string parameter: "human" if you have human genes and "mouse" if you have mouse genes}

\item{plotn}{An integer with the number of term that you want in the plot. The default value is 30.}

\item{url}{A logical parameter. T if you want interactive links}

\item{genes}{A vector with human gene symbol.}
}
\description{
This function is used for RandomComparePheno.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
