% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PhenoEnrichGenespransimu}
\alias{PhenoEnrichGenespransimu}
\title{PhenoEnrichGenespransimu}
\usage{
PhenoEnrichGenespransimu(genes, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with gene symbol, fixed genes or panel genes.}

\item{database}{A string parameter with different options: "ALL", "HUMANDB", "MOUSEDB", "HPO", "CRB" and "MGD"}

\item{organism}{A string parameter: "human" if you have human genes and "mouse" if you have mouse genes}

\item{p.value}{Cut of p value as they are relevant terms}

\item{nulltestnumber}{A number of random genes subset.}

\item{seed}{A seed.}
}
\description{
This function analyses the similarity of phenotypic terms between two groups of genes
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
