% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PhenoEnrichRandomOpt}
\alias{PhenoEnrichRandomOpt}
\title{PhenoEnrichRandomOpt}
\usage{
PhenoEnrichRandomOpt(genes, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with human gene symbol.}

\item{database}{A string parameter with different options. You can define a vector with the databases.}

\item{organism}{A string parameter: "human" if you have human genes and "mouse" if you have mouse genes}

\item{plotn}{An integer with the number of term that you want in the plot. The default value is 30.}
}
\description{
This function makes a plot with enriched phenotype info.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
