% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PhenoGeneNumber}
\alias{PhenoGeneNumber}
\title{PhenoGeneNumber}
\usage{
PhenoGeneNumber(genes, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with human gene symbol.}

\item{database}{A string parameter with different options. You can define a vector with the databases.}

\item{organism}{A string parameter: "human" if you have human genes and "mouse" if you have mouse genes}
}
\description{
This function counts the number of genes per phenotype term from the vector of genes.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
