% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PhenoRelation}
\alias{PhenoRelation}
\title{PhenoRelation}
\usage{
PhenoRelation(enrichresult, genes, p.value = 0.05)
}
\arguments{
\item{genes}{A vector with human gene symbol.}
}
\description{
This function give no overlapping genes
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
