% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PSYGENETloaddercompare}
\alias{PSYGENETloaddercompare}
\title{PSYGENETloaddercompare}
\usage{
PSYGENETloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
PSYGENETloaddercompare
}
