% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{PSYGENETloaddercompareOpt}
\alias{PSYGENETloaddercompareOpt}
\title{PSYGENETloaddercompareOpt}
\usage{
PSYGENETloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
PSYGENETloaddercompareOpt
}
