% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{QueryGenes}
\alias{QueryGenes}
\title{QueryGenes}
\usage{
QueryGenes(genes, database = "HPO", organism = "human")
}
\arguments{
\item{genes}{A vector with the genes symbol}

\item{database}{A string parameter with different options. You can define a vector with the databases.}

\item{organism}{A string parameter: "human" if you have human genes symbol and "mouse" if you have mouse genes symbol}
}
\description{
It produces a data frame with the phenotypical terms from the genes selected and their entrez ids.
}
\details{
This function has developed by Alejandro Cisterna Garcia as part of his PhD mentored by Juan Antonio Botia Blaya
}
