% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{UNIPROTloaddercompare}
\alias{UNIPROTloaddercompare}
\title{UNIPROTloaddercompare}
\usage{
UNIPROTloaddercompare(
  onlygenes,
  genesetcompare,
  organism = "human",
  plotnumber = 30,
  url = F
)
}
\description{
UNIPROTloaddercompare
}
