% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{UNIPROTloaddercompareOpt}
\alias{UNIPROTloaddercompareOpt}
\title{UNIPROTloaddercompareOpt}
\usage{
UNIPROTloaddercompareOpt(
  onlygenes,
  geneset,
  genesetcompare,
  organism = "human",
  plotnumber = 30
)
}
\description{
UNIPROTloaddercompareOpt
}
