% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygenescrb.R
\name{RandomComparePheno}
\alias{RandomComparePheno}
\title{RandomComparePheno}
\usage{
RandomComparePheno(
  geneset,
  genesetcompare,
  organism = "human",
  database = "HPO",
  p.value = 0.05,
  nulltestnumber = 100,
  seed = 1,
  conditional_case = F,
  url = F
)
}
\arguments{
\item{geneset}{A vector with genes symbol.}

\item{genesetcompare}{A vector with gene symbol.}

\item{organism}{A string paramater: "human" if you have human genes and "mouse" if you have mouse genes}

\item{database}{A string paramater with different options. You can define a vector with the databases.}

\item{p.value}{Cut of p value as they are relevant terms}

\item{nulltestnumber}{A number of random genes subset.}

\item{seed}{A seed.}

\item{url}{A logical paramater. T if you want interactive links}
}
\description{
This function measures statistically significant phenotype similarities between gene sets and detects significant differential phenotypes for them.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya
}
